+++ ninja-core/src/test/java/ninja/RouterImplTest.java
@@ -19,6 +19,7 @@ package ninja;
 import com.google.common.base.Optional;
 import com.google.inject.Injector;
 import com.google.inject.Provider;
+import ninja.params.PathParam;
 import ninja.utils.NinjaProperties;
 import org.hamcrest.CoreMatchers;
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -59,6 +60,9 @@ public class RouterImplTest {
         // add route:
         router.GET().route("/testroute").with(TestController.class, "index");
         router.GET().route("/user/{email}/{id: .*}").with(TestController.class, "user");
+
+        router.GET().route("/users").with(UsersController.class, "users");
+        router.GET().route("/users/{username}").with(UsersController.class, "getUser");
 
         router.compileRoutes();
     }
@@ -125,6 +129,80 @@ public class RouterImplTest {
 
     }
 
+    @Test
+    public void testGetReverseGETWithNoContextPathWorks() {
+
+        String contextPath = "";
+        when(ninjaProperties.getContextPath()).thenReturn(contextPath);
+
+        String route = router.getReverseGET(TestController.class);
+
+        assertThat(route, CoreMatchers.equalTo("/testroute"));
+
+    }
+
+    @Test
+    public void testGetReverseGETContextPathWorks() {
+
+        String contextPath = "/myappcontext";
+        when(ninjaProperties.getContextPath()).thenReturn(contextPath);
+
+        String route = router.getReverseGET(TestController.class);
+
+        assertThat(route, equalTo("/myappcontext/testroute"));
+
+    }
+
+    @Test
+    public void testGetReverseGETWithRegexWorks() {
+
+        String contextPath = "";
+        when(ninjaProperties.getContextPath()).thenReturn(contextPath);
+
+        String route = router.getReverseGET(
+                TestController.class,
+                "email",
+                "me@me.com",
+                "id",
+                10000);
+
+        assertThat(route, equalTo("/user/me@me.com/10000"));
+
+    }
+
+    @Test
+    public void testGetReverseGETWithRegexAndQueryParametersWorks() {
+
+        String contextPath = "";
+        when(ninjaProperties.getContextPath()).thenReturn(contextPath);
+
+        String route = router.getReverseGET(
+                TestController.class,
+                "email",
+                "me@me.com",
+                "id",
+                10000,
+                "q",
+                "froglegs");
+
+        assertThat(route, equalTo("/user/me@me.com/10000?q=froglegs"));
+
+    }
+
+    @Test
+    public void testBestFitRoutes() {
+
+        String contextPath = "";
+        when(ninjaProperties.getContextPath()).thenReturn(contextPath);
+
+        String index = router.getReverseGET(UsersController.class);
+        String user = router.getReverseGET(UsersController.class, "username", "james");
+
+        assertThat(index, CoreMatchers.equalTo("/users"));
+        assertThat(user, CoreMatchers.equalTo("/users/james"));
+
+    }
+
     // Just a dummy TestController for mocking...
     public static class TestController {
 
@@ -135,6 +213,35 @@ public class RouterImplTest {
         }
 
         public Result user() {
+
+            return Results.ok();
+
+        }
+
+        public Result getUser(@PathParam("username") String username) {
+
+            return Results.ok();
+
+        }
+
+        public Result users() {
+
+            return Results.ok();
+
+        }
+
+    }
+
+    // Just a dummy UsersController for mocking...
+    public static class UsersController {
+
+        public Result users() {
+
+            return Results.ok();
+
+        }
+
+        public Result getUser(@PathParam("username") String username) {
 
             return Results.ok();